Implement user story 5
In order to get through the barriers
As a customer
I need to touch in and out

DETAILS: This user story is a bit more complex than previous ones. It requires not only the methods touch in and touch out but also a private method that checks whether or not the cars is in journey.